[Heatmap] Add consumer app instruction and fix Heatmap ID
* Heatmap application is designed to have more than one
producer in order to support different types of heatmaps
(runtime-heatmap, trackmap) as well as many heatmap sources.
Consumer will be able to receive all that information in one
place.
* Whereas the manifest templates are already adjusted for that,
the heatmap code was trying to connect to hardcoded 'heatmap:producer'
app. It was needed to be fixed passing HEATMAP_ID param from the
template to the heatmap code.